SCIST/2023-final/CrossStyleSheets/web/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Vincent550102 <5020559@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
